1.3

+++ play/src/main/java/com/star/play/persistence/MovieEntityRepository.java
@@ -2,6 +2,8 @@ package com.star.play.persistence;
 
 import com.star.play.domain.dto.MovieDto;
 import com.star.play.domain.dto.UpdateMovieDto;
+import com.star.play.domain.exception.MovieAlreadyExistsException;
+import com.star.play.domain.exception.MovieDoesNotExistsException;
 import com.star.play.domain.repository.MovieRepository;
 import com.star.play.persistence.crud.CrudMovieEntity;
 import com.star.play.persistence.entity.MovieEntity;
@@ -35,6 +37,10 @@ public class MovieEntityRepository implements MovieRepository {
 
     @Override
     public MovieDto add(MovieDto movieDto) {
+        if(this.crudMovieEntity.findFirstByTitulo(movieDto.title()) != null){
+            throw new MovieAlreadyExistsException(movieDto.title());
+        }
+
        MovieEntity movieEntity = this.movieMapper.toEntity(movieDto);
        movieEntity.setEstado("D");
 
@@ -44,6 +50,9 @@ public class MovieEntityRepository implements MovieRepository {
 
     @Override
     public MovieDto update(long id, UpdateMovieDto updateMovieDto) {
+        if(this.crudMovieEntity.findById(id).isEmpty()){
+            throw new MovieDoesNotExistsException(id);
+        }
         MovieEntity movieEntity = this.crudMovieEntity.findById(id).orElse(null);
 
         if (movieEntity == null) return null;
@@ -53,6 +62,9 @@ public class MovieEntityRepository implements MovieRepository {
 
     @Override
     public void delete(long id) {
+        if(this.crudMovieEntity.findById(id).isEmpty()){
+            throw new MovieDoesNotExistsException(id);
+        }
        this.crudMovieEntity.deleteById(id);
     }
 }
